+++ env.d.ts
@@ -3,6 +3,7 @@
 interface ImportMetaEnv {
   readonly VITE_BACKEND_URL: string;
   readonly VITE_AUTH_URL: string;
+  readonly VITE_AUTH_GUI_URL: string;
   readonly VITE_LOGOUT_URL: string;
   readonly VITE_JWT_URL?: string;
   readonly VITE_METADATA_URL?: string;