+++ src/auth/sbAuth.ts
@@ -6,8 +6,10 @@ import { jwtDecode } from "jwt-decode";
 import { computed, ref } from "vue";
 import { deduplicateRequest, pathJoin, progressiveTimeout } from "@/util";
 import api from "@/api/api";
+import type { User } from "@/store/resource.types";
 
 const AUTH_URL: string = import.meta.env.VITE_AUTH_URL;
+const AUTH_GUI_URL: string = import.meta.env.VITE_AUTH_GUI_URL;
 const LOGOUT_URL: string = import.meta.env.VITE_LOGOUT_URL;
 const JWT_URL: string =
   import.meta.env.VITE_JWT_URL || pathJoin(AUTH_URL, "jwt");
@@ -18,14 +20,15 @@ export type JwtSbPayload = {
   /** Token expiration time as a UNIX timestamp */
   exp: number;
   /** First level keys are resource types, second level keys are resource ids and values are permission levels */
-  scope: Record<string, Record<string, number>>;
+  scope: Record<JwtSbResourceType, Record<string, number>>;
   /** Defines permission levels */
-  levels: {
-    READ: number;
-    WRITE: number;
-    ADMIN: number;
-  };
+  levels: Record<JwtSbLevel, number>;
 };
+
+export type JwtSbResourceType = "corpora" | "lexica" | "metadata" | "other";
+export type JwtSbLevel = "READ" | "WRITE" | "ADMIN";
+
+export const LEVELS: Readonly<JwtSbLevel[]> = ["READ", "WRITE", "ADMIN"];
 
 /**
  * A JWT, if fetched.
@@ -37,9 +40,15 @@ export type JwtSbPayload = {
 export const jwt = ref<string>();
 
 /** JWT payload object with permissions etc. */
-export const payload = computed<JwtSbPayload | undefined>(() =>
-  jwt.value ? decodeJwt(jwt.value) : undefined,
-);
+export const payload = computed<JwtSbPayload | undefined>(() => {
+  if (!jwt.value) return;
+  const payload = decodeJwt(jwt.value);
+  if (!assertValidPayload(payload)) {
+    console.error("JWT payload not valid:", payload);
+    return;
+  }
+  return payload;
+});
 
 /** Return the SB Auth login url, with a redirect back to the given local path. */
 export function getLoginUrl(redirectLocation = "") {
@@ -99,7 +108,7 @@ function assertValidPayload(payload: unknown): payload is JwtSbPayload {
     payload?.scope &&
     "levels" in payload &&
     payload.levels instanceof Object &&
-    ["READ", "WRITE", "ADMIN"].every(
+    LEVELS.every(
       (level) =>
         level in (payload.levels as object) &&
         (payload.levels as Record<string, unknown>)[level],
@@ -111,13 +120,38 @@ function assertValidPayload(payload: unknown): payload is JwtSbPayload {
   return true;
 }
 
-export function hasAccess(
-  payload: JwtSbPayload,
-  resourceType: string,
-  resourceName: string,
-  level: keyof JwtSbPayload["levels"],
-) {
-  assertValidPayload(payload);
-  const scope = payload.scope[resourceType]?.[resourceName];
-  return !!scope && scope >= payload.levels[level];
-}
+/** Get the access level of the current user to a given resource. */
+export const getAccessLevel = (
+  type: JwtSbResourceType,
+  id: string,
+): JwtSbLevel | undefined => {
+  if (!payload.value) return undefined;
+  const scope = payload.value.scope[type]?.[id];
+  if (!scope) return undefined;
+  for (const level in payload.value.levels)
+    if (payload.value.levels[level as JwtSbLevel] == scope)
+      return level as JwtSbLevel;
+};
+
+/** Check if current user has at least READ access to a given resource */
+export const canRead = (type: JwtSbResourceType, id: string): boolean =>
+  !!getAccessLevel(type, id);
+
+/** Check if current user has at least WRITE access to a given resource */
+export const canWrite = (type: JwtSbResourceType, id: string): boolean => {
+  const level = getAccessLevel(type, id);
+  return level == "WRITE" || level == "ADMIN";
+};
+
+/** Check if current user has ADMIN access to a given resource */
+export const canAdmin = (type: JwtSbResourceType, id: string): boolean =>
+  getAccessLevel(type, id) == "ADMIN";
+
+/** Check if a resource user is the currently logged in user. */
+// TODO Identify by idp+sub, not email
+export const isCurrentUser = (other: User): boolean =>
+  other.email == payload.value?.email;
+
+/** Creates the URL to access management for a given resource. */
+export const createAuthGuiUrl = (resourceId: string) =>
+  pathJoin(AUTH_GUI_URL, `resource/${resourceId}`);