+++ src/components/PadButton.vue
@@ -14,7 +14,7 @@ const clickable = computed(() => !!attrs.onClick);
 <template>
   <router-link
     :to="to"
-    class="mink-button w-40 h-40 flex flex-col justify-center items-center text-center hyphens-auto"
+    class="mink-button font-normal w-40 h-40 flex flex-col justify-center items-center text-center hyphens-auto"
     :class="{ 'cursor-pointer': clickable }"
   >
     <slot />