+++ src/corpus/CorpusOverview.vue
@@ -1,5 +1,6 @@
 <script setup lang="ts">
 import { PhPencilSimple } from "@phosphor-icons/vue";
+import SharingPanel from "../auth/SharingPanel.vue";
 import { useCorpus } from "./corpus.composable";
 import useCorpusIdParam from "@/corpus/corpusIdParam.composable";
 import ConfigPanel from "@/corpus/config/ConfigPanel.vue";
@@ -10,6 +11,7 @@ import ExplorePanel from "@/corpus/explore/ExplorePanel.vue";
 import CorpusStateHelp from "@/corpus/CorpusStateHelp.vue";
 import RouteButton from "@/components/RouteButton.vue";
 import LayoutBox from "@/components/LayoutBox.vue";
+import { canWrite } from "@/auth/sbAuth";
 
 const corpusId = useCorpusIdParam();
 const { isConfigValid } = useCorpus(corpusId);
@@ -29,10 +31,19 @@ const { isConfigValid } = useCorpus(corpusId);
             :to="`/library/corpus/${corpusId}/config`"
             :class="{ 'button-primary': !isConfigValid }"
           >
-            <PhPencilSimple weight="bold" class="inline mb-1 mr-1" />
-            {{ $t("edit") }}
+            <template v-if="canWrite('corpora', corpusId)">
+              <PhPencilSimple weight="bold" class="inline mb-1 mr-1" />
+              {{ $t("edit") }}
+            </template>
+            <template v-else>
+              {{ $t("show_more") }}
+            </template>
           </RouteButton>
         </template>
+      </LayoutBox>
+
+      <LayoutBox :title="$t('sharing')">
+        <SharingPanel resourceType="corpora" :resource-id="corpusId" />
       </LayoutBox>
 
       <LayoutBox :title="$t('sources')">