+++ src/metadata/MetadataOverview.vue
@@ -1,6 +1,6 @@
 <script setup lang="ts">
 import { computed } from "vue";
-import { PhTrash } from "@phosphor-icons/vue";
+import { PhLock, PhTrash } from "@phosphor-icons/vue";
 import useMetadata from "@/metadata/metadata.composable";
 import PendingContent from "@/spin/PendingContent.vue";
 import { useResourceStore } from "@/store/resource.store";
@@ -9,6 +9,9 @@ import LayoutBox from "@/components/LayoutBox.vue";
 import RouteButton from "@/components/RouteButton.vue";
 import TextData from "@/components/TextData.vue";
 import FileUpload from "@/components/FileUpload.vue";
+import SharingPanel from "@/auth/SharingPanel.vue";
+import { canAdmin, canWrite } from "@/auth/sbAuth";
+import HelpBox from "@/components/HelpBox.vue";
 
 const resourceStore = useResourceStore();
 const resourceId = useResourceIdParam();
@@ -24,11 +27,12 @@ async function uploadMetadata(files: File[]) {
 
 <template>
   <div v-if="metadata" class="flex flex-wrap gap-4">
-    <div class="flex-1">
+    <div class="w-96 grow flex flex-col gap-4">
       <LayoutBox title="metadata">
         <p>Public id: {{ metadata.publicId }}</p>
 
         <RouteButton
+          :disabled="!canAdmin('corpora', resourceId)"
           class="button-danger"
           :to="`/library/metadata/${resourceId}/delete`"
         >
@@ -36,9 +40,13 @@ async function uploadMetadata(files: File[]) {
           {{ $t("delete") }}
         </RouteButton>
       </LayoutBox>
+
+      <LayoutBox :title="$t('sharing')">
+        <SharingPanel resourceType="corpora" :resource-id />
+      </LayoutBox>
     </div>
 
-    <div class="flex-1">
+    <div class="w-96 grow flex flex-col gap-4">
       <PendingContent :on="`resource/${resourceId}/metadata`">
         <LayoutBox title="content" class="flex-1">
           <TextData
@@ -47,10 +55,15 @@ async function uploadMetadata(files: File[]) {
           ></TextData>
 
           <FileUpload
+            v-if="canWrite('corpora', resourceId)"
             :file-handler="uploadMetadata"
             :primary="!metadata.metadata"
             accept=".yaml,.yml"
           />
+          <HelpBox v-else>
+            <PhLock class="inline mb-0.5 mr-1" />
+            {{ $t("resource.access_denied") }}
+          </HelpBox>
         </LayoutBox>
       </PendingContent>
     </div>